[libc] Fix startup assembly to avoid x18 register clobber
The order of instructions in the inline asm() statement used to warp
the initial thread to its new stacks was susceptible to the compiler
choosing an input operand register such that the initializer for x18
got clobbered.  This happened in the code that GCC generates, by
chance.  So the x18 value meant to be the current shadow-call-stack
pointer was a different address (on the bootstrap stack) such that
Clang-compiled -fsanitize=shadow-call-stack code in the process is
liable to clobber memory and/or crash.

Bug: 3370
Change-Id: I635f06084ad3dfce6baa625dd405b812999dcda4