[alc5663] Add methods for calculating PLL values.
The ALC5663 has a phase locked loop (PLL) which can scale an input clock
signal up or down to a different frequency.

In particular, given an input frequency F_in (in Hz), and three
parameters N, M and K, the PLL generates an output frequency F_out as
follows:

   F_out = (F_in * (N + 2)) / ((M + 2) * (K + 2))

Additionally, the M and K dividers can additionally be bypassed,
removing the "(M + 2)" or "(K + 2)" factors respectively.

This CL adds code to calculate the value of these parameters. We fix
K=2, as recommended by the ALC5663 manual, simplifying the calculation
somewhat.

Change-Id: I6408f5873e16e8fec21038e57637119d44ae9864