[componentmgr] `DestroyChild` actually does deletion
This change makes it so `Realm.DestroyChild` actually follows deletion
through. Before, the "destroyed" realm was kept hidden in a side vector.
The deletion is accomplished by registering the `DeleteChild` action on
the doomed realm.

Part of this change also changes how distinct child instances with the
same name are represented. `ChildMoniker` now distinguishes between
instances with the same name but different instance ids
(`AbsoluteMoniker` will follow). A `PartialMoniker` type is also added,
which is like the old `ChildMoniker` in that it does not distinguish
between instance id.

This change also moves the `Destroy` hook to be invoked upon removal of
the instance from the parent realm. Previously it was invoked after the
instance was "destroyed" but before being removed from its parent.

Change-Id: If05027fd72ffad584d5bf1b9a446ab484f978613