[ledger] Store object identifiers prefixed by digest
This change will allow recovering object identifiers from object
digests.

Updates some PageStorage tests that used to write inline objects
on-disk, which is now failing.

Note: This is a breaking change. If you had Ledger data stored locally
or in the cloud, it will not be readable after this change.

Bug: 35557
Test: fx run-test ledger_tests
Change-Id: I263d7f002eee9db3775a61c335e20a8722ce0d94