[audio_core] Remove latency workaround
This CL removes a short-term workaround latency "pad", and adds
detail to what we display in logs, in case of audio underflow.

Previously, we would "pad" the minimum lead time that we report to
clients by 4ms; this is now returned as-is without padding. Clients
that needed this extra buffering have been updated, so now the
MinLeadTime mechanism can henceforth be used as intended.

Test: unittest for MinLeadTime (from previous CL) has been updated.

Change-Id: I3b3f2ccde81ea6b6d4b0dc659816de524ea43d20