[internet-checksum] Remove unsafe SIMD code and update benchmarks
Recent optimizations made to this crate removed the need for SIMD optimizations so this change
removes that (unsafe) code.

Microbenchmarks were updated to reflect the latest changes to this crate.

SIMD specific tests were removed as the other tests continue to test the existing implementation.

This change also creates a symlink to the root LICENSE file from the root of the internet-checksum
crate's folder.

Tests: Existing tests make sure nothing broke.

Change-Id: If58a4aeb741c5b42beb3319525db01ad644779d5