[zircon][lib][fit] string_view.
Provide a string_view polyfill.
This implementation does not use C++ exceptions, but will assert on required
conditions that otherwise would throw some kind of exception.

The code is written according to C++17 specification, and is compatible
with std::basic_string_view, and the specializations (char, wchar, etc).

Currently only basic_string_view<char> is provided as a public type.
Though other specializations can be provided if and when necessary.

Tests are written to provide two types of checks, adherence to the
standard and correctness. This is check that the expected result is
correct, and the requirements of equivalency defined in the standard are
mantained.

Required for: https://fuchsia-review.googlesource.com/c/fuchsia/+/312468

Reference: https://en.cppreference.com/w/cpp/string/basic_string_view

TEST=fit-test

Change-Id: Iaba7c1e2c49b5487cd5c9f9bd491e8893fe77e32