[escher] Immediately reset new timestamp query pools.
Queries must be reset before first use, not just between
uses.  A recent update to the Vulkan validation layers
notified us that we weren't doing this.

The validation layers also notified us that Scenic is
always making timer queries, which should only happen
during tracing.  This is also fixed, by respecting the
"enable_gpu_logging" arg passed into the escher::Frame
constructor.

Finally, we reduce the size of the query-pools from 100
to 20; this is still more than we use.

Bug: 35588
Bug: 35627

Change-Id: Ifeb13e60b245537c0626456874b531d8b007ff12