[fdio] Add fdio_open_fd{,_at}
This is a convenience wrapper around fdio_open() and fdio_fd_create().

For programs that wish to spawn child processes, we wish to require that
they call fdio_get_vmo_exec().  That function takes a file descriptor,
and that file descriptor has to be opened with
fuchsia.io.OPEN_RIGHT_EXECUTABLE, which we do not expose to the Unix
compatibility layer, so such programs would otherwise need to create a
channel and then bind it to an fd, both of which are implementation
details the client doesn't particularly care about.  This will make the
client code for spawning a child process a bit simpler.

Bug: 35646
Change-Id: I4782d2edad550b4d9f102b501301461f055fbb73