[mediaplayer] use sysmem for mediacodec decoding
This CL updates the node implementation that wraps mediacodec so it
uses the new sysmem payload mode. This means sysmem is used to
allocate buffers for both the input and the output of the decoder
when a mediacodec decoder is used.

Most changes involve the BufferSet class, which is used to manage
buffers for both the input and the ouput of the decoder. A new unit
test is added to prevent BufferSet regressions.

TEST: new unit test for BufferSet

Change-Id: I30abdb849cb08f23cc79da119bf5e2126473b9fa