[netstack3][ndp][slaac] Do stateless global address autoconfiguration for IPv6
Perform SLAAC for global addresses as defined in RFC 4862 section 5.5.

Tests: Unittest to make sure an address is generated and lives only for its specified
lifetime. Tests to make sure an address is NOT generated if the prefix option is
invalid/malformed. Test to make sure that only hosts perform SLAAC for global addresses.

Change-Id: Iab5f848d1e81c86ada20b057717f6449a1ccf3af