[benchmarks] Use #if to conditionalise instead of builder/bot name
The motivation for doing this is that I want to add these tests to get
run on the perfcompare trybot.  That has a different builder name from
the peridot-perf builder that these tests currently run on, and I
don't want to have to check for extra builder names or parse the
names.  Using an #if is simpler.

The main reason this code uses the builder name rather than an #if is
that this code used to be a shell script, where there wasn't a way to
conditionalise on architecture.

Currently this benchmarks.cc file only runs on NUCs.  There are VIM2
hardware builders but they are currently not working.  So the
conditionalisation can't be fully tested, but we'll keep it in case
the VIM2 builders start working again.

Bug: 35472
Test: peridot-perf trybot
Change-Id: I24b0d61a551b00d3df394a01aba958d463bd0415