[fidlcat] start object syscall implementation
- zx_object_wait_one
- zx_object_wait_many
- zx_object_wait_async
- zx_object_signal
- zx_object_signal_peer
- zx_object_get_child

Allow a memory buffer to be loaded twice
- once for an input argument
- once for an output argument

This is used by zx_object_wait_many which has items
which is an input/output array.

If a buffer is only used by an input argument or an output argument, the
buffer is still loaded only once.

Allow interception tests to modify a buffer during the
syscall simulation.

Change-Id: Ic94998bfb397639dad0bb4978324e6fdeb430067