[ledger] Add background sync of closed pages.
BackgroundSyncMamager is responsible for starting synchronization of closed pages with the cloud. The sync is triggered every time BackgroundSyncManager is notified about a page being unused (has no external nor internal connections).
Current implementation is expected to trigger the sync of the page repeatedly, even if it has just being closed after background sync, as there is no mechanism to track the sync state of pages.

Change-Id: Id50586806800da4c824c222ec0480502a9ce5396