[conn][packet] Fragmented serialization
- Updates trait hierarchy in packet crate to allow for fragmented
buffers to be used in serialization.
- Adds FragmentedBuffer type as the base type.
- Adds ShrinkBuffer, GrowBuffer and GrowBufferMut traits, which allow
tighter trait bounds based on the needed operations.
- Adds TargetBuffer trait, which is a buffer type that can be serialized
into.
- Adds ContiguousBuffer trait, a marker trait identifying that the
underlying buffer is a contiguous byte slice.
- ParseBuffer and ParseBufferMut are bound to Shrink + Contiguous.
- ReusableBuffer is introduced as any buffer that is TargetBuffer +
Shrink, allowing tight bounds on the blanket implementations of
BufferProvider and BufferAlloc.
- Buffer is now just a ParseBuffer that can grow.
- BufferMut is a ParseBufferMut that can be a serialization target.

- The SerializeBuffer API is kept artificially the same with unsafe
joining and possible panics. That will be removed in a follow-up CL as
downstream dependencies of the packet crate are updated to the intended
final API.

TEST: No behavior change is expected from this reorganization, as long
as the used buffers are contiguous (which is the case, since it was not
possible to create non-contiguous buffers). Tests effectively
demonstrating a scatter-gather buffer in use are in a follow-up CL,
along with the new API for SerializeBuffer.

Change-Id: I68b1a6ad9c2c56191cd0639f15f98a529a02ff8f