[netstack3] Reference count joined multicast address groups
This change addresses a bug where if the same multicast group is joined twice, it will be left
after the first attempt to leave the multicast group. This is an issue for IPv6 as many addresses
may map to a single solicited-node multicast address, and leaving any single address will leave the
same solicited-node multicast group of (potentially) many other addresses. This can cause missed
neighbor solicitation messages which will result in duplicate addresses on the network, which will
almost definitely cause problems within the link.

Tests: Unittest to make sure that the multicast address is only left after multicast addresses'
counter reaches 0.

Change-Id: I7523535951ac09ebd1fd4e75ecf7a652eb99b1e2