[pkgfs] Fidl api service definition for pkgfs for corrupted blobfs.
This change adds a fidl protocol definition for blobfs corruption
indication. Due to dependency of pkgfs on blobfs,
blobfs initially behaves like a server to get a handle from pkgfs,
which is later used by blobfs to send merklehash of corrupted
blobfs to pkgfs.

So pre-pkgfs init phase the behaviour is same as before i.e
read and return error to clients.

Post pkgfs init stage, read corrupt blobs will lead to
calling pkgfs with the corrupted blob merkleroot.
Depending on the response, blobfs will either delete the blob and hang or
delete the blob and return the error to the client.

Fixes: ZX-3139

Testing: Not needed. Simply adds a fidl api defination.

Change-Id: I9984b4bf49b26e47609f5ca278854b715caa88a8