[DriverFramework-DDK] Use the latest suspend_new hook when available
This change modifies the devhost functions to use new suspend hook
when drivers have them defined. The old suspend hook takes the suspend
flags as input, while the new suspend hook takes the desired device
power state as input. The devhost converts the suspend flags to the
device power state and calls the suspend_new hook with that state.

Test: 1.Added automated tests to see the suspend_new hook gets called,
when defined. Verified that the old suspend hook is called if a driver
does not define it. runtests -t ddk-power-test
2. dm reboot on qemu.
3. dm reboot on Astro[in progress]

Change-Id: I1d8dad3cf751535ff274acbb0d6ffd535bba46e2