[netstack] Fix ARP timeout error handling on socket writes.
netstack was silently shutting down write operation on a socket
when any transmit over the socket experiences ARP timeout.
We should not be shutting down r/w on the socket, but just drop
the write and continue. A new error stat for arp timeouts would be
added to gvisor netstack.

Bug: b/138159692
Test: fx run-test netstack_c_api_tests -t netstack_external_network_test \
	  -- --test.run=ExternalNetworkTest.UDPErrSend
Change-Id: I2e9e6ebe9ab33254aa6075dac578ebcff986daba