[benchmarks] Use ".json" extension consistently for all the perf results files
This is preparation for adding more tests to get run on the
perfcompare trybot.

perfcompare.py currently looks for files with a ".json" extension, but
some of the results files (those with names generated with
MakeTempFile()) did not have any filename extension.

This change makes the generation of filenames more uniform so that:

 * All the filenames are guaranteed to be unique.

 * All of the filenames have an informative name (not just
   "benchmarking_temp_file_N").  That's useful if you need to look
   through the output directory.

Also use the fuller extension ".fuchsiaperf.json" for consistency with
sl4f, which was changed earlier here:
https://fuchsia-review.googlesource.com/c/fuchsia/+/301117

Bug: 35472
Test: peridot-perf and fuchsia-perfcompare trybots
Change-Id: I49c6eeec2502c65143b4e699d8a62d23f327d45e