[componentmgr] Embed instance ids in monikers
Instance ids are used to distinguish component instances with the same
name. This can happen if an instance is deleted and a new one is created
with the same name. This is important to distinguish isolated storage
between like-named component instances, for example.

They are already present in ChildMoniker, so here we add them to
RelativeMoniker and AbsoluteMoniker.

Change-Id: I7afb70fffd4d0b9436e4c614c34aa622600bc9d1