[sync][completion] Optimize completion's use of futex_wake
Update the libsync completion construct to keep track of whether or not it has
waiters while in the unsignaled state.  This allows the implementation to call
futex_wake only when there can be waiters, instead of unconditionally.  In
addition to reducing the number of syscalls which have to be made, this also can
reduce the potential for contention over the futex contex lock in the kernel.

See details in ZX-4734 for a more detailed analysis of where this might matter.

In addition, during this optimization, a potential race condition was
identified by inspection.  It has been corrected, but this introduces another
potential race.  They are as follows.

Race 1:
1) Thread A waits on an unsignaled completion X.  It fully joins the futex wait
   queue.
2) Thread B signals X and futex_wakes A.
3) Before Thread A returns from futex_wait, thread B (or C, or whatever)
   unsignals X.
4) Thread A wakes.  Its return code from futex_wait is OK.  The old code
   interpreted this code to indicate that it should cycle through the
   completion_wait path all over again.  This time, it sees that the futex is
   unsignaled, and ends up waiting again.

The net effect of this is that it was technically possible (with a bunch of user
flapping on the signaled/unsignaled state of the completion) to cause a thread
to miss a wake event.  This has been corrected.

Race 2:
1) Completion X is unsignaled has a set of current waiters, {N}.
2) Thread A signals the futex, but after transitioning the internal atomic state
   to SIGNALED, but before it calls into futex_wake, the following operations
   occur.
3) Thread B unsignals the futex.
4) Thread C waits on X and fully joins the futex wait queue.
5) Thread A makes its call to futex_wake.  The set of threads which was woken is
   {N} + C instead of being just {N} as it should be.

This issue is because C no longer checks the for the unsignaled state of the
completion.  In fact, if Race #1 were still present, {N} + C would remain
blocked instead of {N} + C being released.

In the first case, it is possible to stall a thread which should be woken.  In
the second, it is possible to spuriously wake a thread which should remain
blocked.  Between these two choices, I've decided that the spurious wakeup is
more likely to be tolerated (given typical usage patterns) than the occasional
improper stall.

This whole situation can be fixed by introducing a generation counter to the
signaled/unsignaled state.  For now, the increased complexity that this would
result in does not seem worth it.  Having multiple threads all hammering on a
completion event with signal/unsignal/wait operations all at the same time seems
like a very unusual usage pattern already; and any user of such a pattern had
better be prepared for some significantly racy behavior already (eg; they should
already be hardened against spurious wakeup)

34507 #comment Implemented the condition_t mitigation.

Tests: Existing unit tests, also booted and ran a workstation build and verified
the reduction in futex_wake operations via tracing (bug 34507)

Change-Id: I68d35c6d13478710d17d6c01f06999e7ba3b1173