[alc5663] Enable output audio path.
We enable the output audio path on system boot, and leave it in "always
on" mode. This is simlar to other codecs for Intel HDA SST. Longer term,
we will add support for dynamically configuring / powering the codec.

In "alc5663_registers.h", some register definitions are marked with
"(??)". These are definitions that are not referenced in the ALC5663
datasheet, but are nevertheless required for audio output to work.

The values were derived by comparing the I2C register values of a live
ALC5663 codec playing audio on a working system against the values
specified by the ALC5663 datasheets. The names and field locations of
such registers are guesses based on inspecting datasheets of other
Realtek codecs, and empirically trying different values.

Bug: 35648
Change-Id: Ia23a908c2e5416836cf67a18bfe323ac9702652c