[inspect] Replace GetVmo
Instead of providing access to the underlying VMO using GetVmo, this
change replaces that method with new DuplicateVmo and CopyVmo methods.

DuplicateVmo returns a read-only duplicate of the underlying VMO that is
ready to be send to a reader.

CopyVmo returns a copy of the underlying VMO data, truncated to the last
in-use page. This data is ready to persist to disk.

Change-Id: I08c3f55ee884a259d844239437bc39b86c784814