[prebuilt_test_manifest] Change run_test_component by run.
By default we were using run-test-component to run prebuilt tests. It
turns out that some tests need to use the real fiddle services and use
"run" instead of "run-test-component".

With this change we are providing "run" as the default runner and
provide functionality for prebuilt test owners to optionally provide
a different runner for their tests.

Test: Run prebuilt tests manually using run.
Bug: 10331
Change-Id: I9fc97d8c6a7eea1f240211dfa2fbef5a6bad84bd