[fdio][spawn] Move #! handling into fdio_spawn
This moves the handling for shebang (#!) directives out of launchpad and
into fdio_spawn. In other words, script interpreters are loaded on the
client side now rather than in the fuchsia.process/Launcher service.

One significant behavior change also included in this CL is changing
script interpreters to be loaded from the calling process' namespace
instead of through fuchsia.ldsvc/Loader.LoadScriptInterpreter. Besides
the fact that this is simpler, this is also done because it was
discovered that LoadScriptInterpreter is currently broken and allows
processes to escape their namespace sandbox. (See fxb/35818)

Primarily, this unblocks moving the system off a launchpad-based
fuchsia.process/Launcher to the new Rust process_builder-based Launcher
hosted in component_manager.

Bug: 25077
Bug: 35818
Change-Id: Ifa68535fde268c7bfb84534a2035adfce015d4ca