[cml] expose:to
CML schema addition: `{"expose":[{"to": ("realm"|"framework")}]}`

This change does not implement handling of this addition, but it adds
"to" to the appropriate part of the schema, plumbs this info to the
layer consumed by `component_manager`, and updates affected test code.

This change is towards WorkScheduler's `Worker` interface, which should
be exposed with `"to":"framework"`, allowing `component_manager` to
invoke scheduled units of work via this interface.

Bug: 35215

Change-Id: Ide986faab107b00f94cdf695914a9bf7cbde2418