[fs-management] negative fsck test corrupts inode bitmap
this test was originally corrupting the superblock, but there are
changes in the pipeline for adding a backup superblock which means that
fsck actually succeeds in this case. instead we corrupt the inode
bitmap, which doesn't have a backup. the inode bitmap is at a stable
location right now, but ideally we should parse it's location out of the
superblock to be resilient to future changes.

Test: fx run-test fs-management-test
Change-Id: Idd189072797feffd71260620ebf7fdd18a6bf0f6