[bt][a2dp] Fix player buffer not receiving audio
Previously we used a child VMO set to COW for the playback buffer. This
wasn't necessary and broke when the COW behavior changed to that
children didn't mirror modified data in the source.  Just use a
duplicated handle to the same VMO instead.

Also updated some of the code around setting up the buffers, and require
that an SBC player be present as required by the spec.

Bug: 2806
Test: fx run-test bt-a2dp-sink-tests
      added test_correct_frames_are_sent

Change-Id: If1f496b7401fafb8af3b1e4860c37a1e478b4514