[scripts] fx sniff
fx sniff captures the packets flowing in-and-out the Fuchsia target
device, and displays the packets in a useful format. This is to run
on the development host.

fx sniff automatically filter out fx-workflow related packets, such
as ssh, package server, logs, zxdb, etc., so that the user can focus
on the application of interest. For those who need to debug fx workflow
itself, the full command under use is also printed; one can easily
modify to meet their own needs.

[Typical usages]
$ fx sniff --help                  # show all command line options
$ fx sniff wlan                    # capture packets over WLAN,
                                   # and show their summaries
$ fx sniff --format hex wlan       # hex dump
$ fx sniff --format wireshark wlan # capture packets over WLAN,
                                   # and start wireshark GUI for display
$ fx sniff -t 10 --file myfile eth # capture packets for 10 sec and
                                   # store at //out/myfile.pcapng

Change-Id: I39c749a7e0cfff0ef7f8f962b71ff92cccbc71af