[sherlock][thermal] Disable Thermal Driver
When the driver was ported from Astro to Sherlock
it was missed that the CPU arch is diffferent.
Astro has a single cluster, but Sherlock has two clusters
So this existing driver needs support for that.

As of today, this driver sets the A73 freq but A53 voltage,
so disabling this driver makes sense since we will be enabling
A73 cores again.

Test: No testing needed, disabling driver at top level

Change-Id: I23f6001333b150996911e8a9f98fc60176d162e8