[blackout] add minfs power failure test
Add a simple minfs power failure test to blackout. This test formats
minfs, creates a file, writes to it, closes it, opens it, reads from it,
verifies it's contents, closes it, deletes it, then does it all again,
forever. Future work includes generating more interesting and
sophisticated loads on the filesystem. It reboots the system using dm
reboot, although since the filesystem is launched by our process it's
not cleanly shutdown as part of the normal shutdown process. During the
verification step it runs fsck, and does no additional verification.

BUG: 35789

Change-Id: I60b618c1900b04410d9bece9df8b4a60bbe21d77