[perfcompare] Handle tests that produce just a single value per process run
This is preparation for adding more tests to get run on the
perfcompare trybot.

Currently all of the tests run with perfcompare use the libperftest
library.  Those all do multiple test runs per process run.  However,
some other tests produce just a single value per process run.  For
example:

  {
    "label": "scenic_total_frames_drawn",
    "test_suite": "fuchsia.scenic.hello_scenic",
    "unit": "count",
    "values": [585]
  }

For now, change the behaviour to drop the initial value only if
there's >1 value.  In the future we might want to change that
threshold to be higher than 1, or change the behaviour based on the
"unit" field, but that doesn't seem necessary yet.

Also update the issue reference from Jira to Monorail.

Bug: 35472
Test: "python perfcompare_test.py" (run by bots on CQ) - test case added
Change-Id: I6974748c1fdc146b58fdf2cb17bfdc28f8de9937