[netstack] filter: Support a range of ports in a rule
This CL replaces every `port` in the rules with `port_range`, which
has `start` (uint16) and `end` (uint16) fields. The range is
inclusive. If `start` and `end` are equal, it represents a single
port. If `start` and `end` are 0, it represents "any" port.

In a text rule, you can still write `port N` but you can also
write `range N:M` now.

Bug: 21175
Change-Id: Id3fd098b3cba608555f9d634333af1c209613d1e