[kernel][x86] Avoid no-op page table updates
When updating a mapping entry this now checks if the entry to be
installed is the same as the existing one. If it is we can leave it and
not have to perform cache flushes or tlb invalidations.

The use case here is where you want to protect a range of mappings that
may or may not need changing. The alternative would be to first query
all the mappings and then work out the ones that should have protect
called on them. Currently there is no way to efficiently query a range
of mappings.

Change-Id: I84a00e33f8b2a1da71d9400a42385fcea715b7ef