[sdk][build]: Fix build by removing un-necessary dependencies.
My build was broken this morning after a jiri update, and
this CL fixes the issue (conformance_test.cc not compiling)
by removing two un-necessary dependencies from the
//sdk/lib/fidl/cpp:conformance_test_sources GN target.

One of them (:fidl_test) was inserting an include search
path in the compiler invokation, which make the source
file include the wrong header file (see bug for details).

Since the source file doesn't use any :fidl_test header
or feature at all, it can be removed, solving the issue.
The :cpp_base target is also not needed and thus removed
from the list of direct dependencies (even though it
might be an indirect dependency).

BUG: 35936
TEST: "fx build" rebuilds everything correctly.
R=bprosnitz@google.com

Change-Id: I3d2e8b72d46d0b01d3babb7ebf82f153dbe400ab