[pkg-resolver] Add hyper based update source test
This change adds a new helper to the Repository testing helper to serve
the repository via HTTP without having to spawn a 'pm serve' process.
This server exists in fuchsia_pkg_testing for now as it only serves the
static files in the repo and does not yet handle /config.json or /auto.

Bug: 22818
Change-Id: I711b500dd64783f8b148d06eee4995ae0f31c648