[kernel][wait] Assert if we attempt to block while holding a spin lock.
Any time we enter a wait queue in order to block we should immediately
assert if we are holding any spin lock aside from the thread_lock.
The thread lock is required and will be logically transferred to the
thread who takes our place on this core after we block.

Change-Id: I01df90e3d34807cc4037f34b9608845c18a2791c