[component_mgr] Implement shutdown capability
Component manager exposes the fuchsia.sys2.SystemController
protocol. The implementation inside component manager when
SystemController.Shutdown is called is to simply exit the
process. This primitive behavior will be replaced with a more
sophisticated and orderly behavior later.

This change also adds in integration test that starts component
manager as a v1 component. The root component of the component
manager under test has a component which consumes the
SystemController capability and calls Shutdown.

TEST=shutdown_integration_test

Change-Id: Ie73993253ba6db39fcbd5b36e22a6c1ff228cbdf