[a11y] Implements a11y Gesture Arena.
This change implements a Gesture Arena to be used by accessibility services
inside accessibility manager.

It enables recognizers (which extend the GestureRecognizer class), to join the
arena and contend for the current gesture being performed on the touch screen.

Recognizers either win or lose the current arena contend, which helps to
disambiguate gestures.

This arena is heavily influenced by Flutter's gesture arena, although it makes
some simplifications which do not affect accessibility services requirements. Of
important notice is the lack of a sweep mechanism, which clears an arena if no
gesture recognizer has responded after an interaction is complete. This
simplification is possible because at the moment we own every recognizer that
will be implemented, so we can enforce that they will always respond either as
winners or not.

This change also introduces the possibility of individual recognizers to consume
/ reject streams of pointer events, which is a functionality described in
|fuchsia.ui.input.accessibility| protocols.

Finally, with this change, the goal is to add a way for individual recognizers
to be able to add semantic meaning to sequences of raw pointer events in the
form of gestures as well as the possibility to decide where those pointer events
will go.

TEST:
    fx run-test a11y_lib_tests -t gesture_arena_tests

fxb/35406 #comment

Change-Id: I9b0fc0001d9e57293944ce18db51eb8f49bacb2f