[machina] Fixes guest_interaction_tests.
Fixes a bug in client_operation_test where an event could go unhandled
until the test case was torn down.  The implementation of the callback
for the event is destroyed before the tear down method runs, so when
the events were flushed during tear down, the test could crash.  This
change waits for the final event to be triggered before ending the
test.

Bug: 35954
Test: for i in {1..100}; do
    fx run-test guest_interaction_tests -t client_operation_test
  done
Change-Id: I4f1418fa4375eeaa50ccdd9d138a1a859d175cbe