[media][drm] Add DRM error codes
This CL adds two additional error codes to fuchsia.media.drm.Error:
INTERNAL and NOT_PROVISIONED. The former is a catchall error code for
non-actionable DRM system specific errors. The latter is to indicate
to the client that the server lacks the provisioning necessary to
acquire licenses. If the DRM system allows for dynamic provisioning, the
client can do so using the fuchsia.media.drm.Provisioner protocol.

Change-Id: I45a4b6d772a5389cd79006e4b80d76768d6ceb21