[audio_core][effects] Remove 'unloaded' state from EffectsLoader.
EffectsLoaders can now only exist in the 'loaded' state, moving the
logic that was previously in 'LoadLibrary' into static factory methods.
We also provide a 'null' factory method that can be used to provide a
no-op EffectsLoader that can be used if no actual library exists. The
null loader is simply a loader that knows about 0 effects.

Test: fx run-test audio_effects_loader_unittests
      fx run-test audio_core_unittests
      fx run-test audio_effects_example_tests
Change-Id: I997324b7c766852ecb41c8fa3b4ee6aace970037