[perfcompare] Display units of measurement in the output results table
For example, instead of "2103 +/- 43", display "2103 +/- 43 ns".

When the perfcompare trybot ran only the microbenchmarks, the units
were always nanoseconds, so including them wasn't essential but would
be useful, since it's more explicit and makes the output more
self-documenting.

Now that the perfcompare trybot runs other tests, it's more important
to include the units because some tests use different units, such as
"ms" (milliseconds) for the .tspec-based tests.

Bug: 35472
Test: "python perfcompare_test.py" (run by bots on CQ) + perfcompare trybot
Change-Id: Ifaa6cce2374d03278db1a6bb2d9b2705edd5ddf4