[mediaplayer] replace 'decoder' with more general 'processor'
This CL changes many uses of the term 'decoder' to the more general
term 'processor' in anticipation of decryptors. Decryptors will share
the Processor class definition as well as the FidlProcessor
implementation, which wraps a fuchsia.media.StreamProcessor.

DecoderFactory is not renamed as it still creates decoders.

The Create function and constructor for FidlProcessor omits some
unneeded parameters and now takes medium (audio/video/etc) and
function (decode/decrypt/etc), primarily for debugging purposes.

This CL is largely a name change. Current tests are adequate for
detecting regressions.

TEST: no behavior change
Change-Id: I72c689628f24b6d80609f6a7e7bf251879f9484f