[netstack3] Add `should_panic!` macro, make panic tests more precise
This change adds the `should_panic!` macro. It behaves similarly
to the `#[should_panic]` attribute on tests, but instead of requiring
that any line of code in the test panics with a given message, it is
used to wrap a particular line of code to ensure that that line in
particular is the one that panics.

This change transitions two existing tests to use the macro; more
tests will be transitioned later.

Test: Transitioned tests in the wire::udp::tests module

Bug: 35927

Change-Id: I6ccb0e9ccabc57719a88b5e1a3412771221c6ad9