[hwreg] Remove superfluous "const" in DEF_ENUM_SUBFIELD definition.
The definition of DEF_ENUM_SUBFIELD() has an unnecessary "const" in it,
causing GCC to issue the warning:

  error: type qualifiers ignored on cast result type [-Werror=ignored-qualifiers]
         .set(static_cast<const typename std::remove_reference<decltype(FIELD)>::type>(val));

This is not caught by existing tests, as the warning (for unclear
reasons) does not appear to be generated when the structs are defined
inside the scope of a function.

Remove the redundant "const", and add a test case exercising globally
scoped structs containing subfields.

Change-Id: I3bcb2a997d2080c483c557b59fb95a54cc18b73b