[alc5663] Set up device clocks.
We need to configure the ALC5663 to have a clock of 256*|sample_rate| for
its system clock.

The ALC5663 gives us the choice of using MCLK or BCLK. We don't (yet)
have a way for the SoC to communicate MCLK to us, so we currently choose
to use BCLK as our main clock.

Because |bclk_frequency| won't be high enough to provide
a 256*|sample_rate| clock, we need to plumb it into a PLL to get the
right frequency, resulting in the following timing chain:

   (BCLK) ---> (Clock select) ---> (PLL) ---> (Clock select) --.
                                                               |
           (System clock) <---clk_sys_i2s <--- (Divider) <-----'

Once the I2S BCLK gets plumbed through the PLL, the resulting clock isn't
synchronized with original clock any longer. When the clocks are not
synchronized, the ALC5663 requires us to enable the "asynchronous
sampling rate converter" (ASRC).

The datasheet is a little unclear about what the system clock needs to
be when ASRC is enabled. Section 7.5 suggests the system clock should be
`256*|sample_rate|`. Section 7.5.2 suggests when ASRC is enabled, the
system clock must be at least `512*|sample_rate`. Emperically, it
appears that the clock `clk_sys_pre` needs to be at least
`512*|sample_rate|`, while the clock labelled `clk_sys_i2s` needs to be
`256*|sample_rate|`. Confusingly, both are called the "system clock" in
different parts of the manual.

Tests: Instead of just verifying that registers are set to a fixed set
of values, we attempt to re-calculate from scratch the clocks in the
same way the hardware would, and ensure they are in range.

BUG: 35648
Change-Id: I2d10e8b8edf81b551cfe4326b2b45f7f1756955e