[kernel][vm] Use protect to downgrade mappings
Use arch protect to downgrade mappings from writable to readable. See
bug for extended performance results.

Presently protect is either as fast or faster as unmap and so it
makes sense to use it if the mappings will be touched again. If the
mappings are not touched then they still have to be unmapped later
and the time spent doing the protect was wasted.

Test: Added a new test that exercises this code via creating creating
VMO clones and validates mappings are protected correctly.

Bug: 35035

Change-Id: I933d366fd0e3010665ee9feee5c7911b9fcc3125