[kernel] Fix multi Alloc+Free race in GPArena
The GPArena was vulnerable to an ABA race where if
 * thread A was in Alloc and had already read head->next but had not
   yet performed the CAS
 * thread B races and allocates head, head->next and then free's head
 * thread A will then successfully complete the CAS, since head is
   unchanged
Unfortunately the value of head->next it had read is stale and points
to a now allocated object. Ultimately this leads to performing double
allocations, which naturally results in corruption.

Solution chosen here is to pack the head and head->next values together
and operate on them using a 16byte CAS. This ensures that if any such
race occurs thread A will now know that they have a stale head->next,
as the CAS will fail.

An alternative, and more classical, solution would be to use the low
bit of the pointer as a flag to indicate that the node is being
operated on. This effectively 'locks' the node until the thread has
a chance to complete its removal. Whilst it has the advantage of
avoiding the need for 16byte CAS, it also greatly complicates the
algorithm.

Test: The underlying race condition being fixed here can be caught by
the existing GPArena tests, although it triggers very very rarely. I
was unsuccessful in writing any better tests that had a reasonable
trigger rate.

Bug: 34522

Change-Id: I0658fb39ab907d9d3be1eedd2a9db14e5bff6375