[kernel] User copy variants that capture and return faults
The additional user copy variants allow for the caller to decide what
to do if a page fault occurs. Intended use case is for where the caller
needs to drop locks being held so the page fault can be handled.

At the arch level this is implemented by using the high bit of the
fault return pointer to indicate whether the return should be done
before or after calling the page fault handler.

Bug: 34639

Change-Id: I01f68de75fdd47fff8cf625a2bb19ef849441a61